% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/density_at.R
\docType{methods}
\name{density_at}
\alias{density_at}
\alias{density_at,gaussian_dist,matrix-method}
\title{Density At}
\usage{
density_at(omega, x)

\S4method{density_at}{gaussian_dist,matrix}(omega, x)
}
\arguments{
\item{omega}{The distribution (p-dimensional)}

\item{x}{An n x p matrix of points}
}
\description{
Density At
}

