% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gauss_class_initialize.R
\docType{methods}
\name{initialize,gaussian_dist-method}
\alias{initialize,gaussian_dist-method}
\title{Initialize the gaussian_dist class}
\usage{
\S4method{initialize}{gaussian_dist}(.Object, covariance, ...)
}
\description{
Initialize the gaussian_dist class
}
\seealso{
\code{\link{initialize}}
}
\keyword{internal}

