% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gauss_class_initialize.R
\name{new_gaussian}
\alias{new_gaussian}
\title{Create an instance of the gaussian dist}
\usage{
new_gaussian(covariance)
}
\arguments{
\item{covariance}{The covariance matrix}
}
\description{
Create an instance of the gaussian dist
}
\examples{
cm <- diag(rep(1, 3))
gd <- new_gaussian(cm)
x <- sample_points(gd, 10)
de <- density_at(gd, x)
}

