% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gauss_class_initialize.R
\name{new_simulation_pars}
\alias{new_simulation_pars}
\title{Create an instance of the simulation params}
\usage{
new_simulation_pars(prior, sigma, k)
}
\arguments{
\item{prior}{The distribution}

\item{sigma}{The noise level of observed points (square root of sigma^2)}

\item{k}{The number of classes}
}
\description{
Create an instance of the simulation params
}

