% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gauss_class_initialize.R
\name{run_simulation}
\alias{run_simulation}
\title{Run a simulation}
\usage{
run_simulation(params, external_its, internal_its)
}
\arguments{
\item{params}{The simulation meta-parameters}

\item{external_its}{The number of outer loops where parameters are randomized}

\item{internal_its}{The number of inner loops where random variables are drawn}
}
\description{
Run a simulation
}
\details{
\code{id_rates}
A vector of identification rates per outer loop
\code{id_rate}
The average across all trials
}

