% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gauss_class.R
\docType{methods}
\name{sample_points}
\alias{sample_points}
\alias{sample_points,gaussian_dist,numeric-method}
\title{Sample Points}
\usage{
sample_points(omega, n, ...)

\S4method{sample_points}{gaussian_dist,numeric}(omega, n)
}
\description{
Sample Points
}

