% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/theory.R
\name{theoretical_rate}
\alias{theoretical_rate}
\title{theoretical_rate}
\usage{
theoretical_rate(params, de = density_at(params@prior,
  sample_points(params@prior, 1e+05)))
}
\arguments{
\item{params}{Simulation parameters}

\item{de}{A sample of densities for points drawn from the prior}
}
\description{
theoretical_rate
}

