% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gauss_class_initialize.R
\name{new_gaussian}
\alias{new_gaussian}
\title{Create an instance of the gaussian dist}
\usage{
new_gaussian(covriance)
}
\arguments{
\item{covariance}{The covariance matrix}
}
\description{
Create an instance of the gaussian dist
}

