% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sample_points.R
\docType{methods}
\name{sample_points}
\alias{sample_points}
\alias{sample_points,gaussian_dist,numeric-method}
\title{Sample Points}
\usage{
sample_points(omega, n)

\S4method{sample_points}{gaussian_dist,numeric}(omega, n)
}
\arguments{
\item{omega}{The distribution from which to sample}

\item{n}{The number of observations to sample}
}
\description{
Sample Points
}

